Move Resolution class to its own file
